+++ BTCode/Speciator.cpp
@@ -24,13 +24,13 @@ void Speciator::sortSpecie(Specie& spec) {
 }
 
 void Speciator::fitnessSharing(std::vector<Tree>& trees) {
-    for (int i = 0; i < trees.size(); i++)
+    for (size_t i = 0; i < trees.size(); i++)
         adjustFitnessShared(trees, i);
 }
 
 void Speciator::adjustFitnessShared(std::vector<Tree>& trees, int index) {
     int divisor = 1;
-    for (int i = 0; i < trees.size(); i++) {
+    for (size_t i = 0; i < trees.size(); i++) {
         if (i != index && sameSpecie(trees[i], trees[index]))
             divisor++;
     }
